// services/api/core/models/interface/IApiRequest.ts
import type {EApiResponseType} from "~/services/api/core/models/enum/EApiResponseType";

export interface IApiRequest {
    method: string;
    endpoint: string;
    data?: any;
    queryParams?: Record<string, any>;
    pathParams?: Record<string, any>;
    headers?: HeadersInit;
    responseType?: EApiResponseType;
    cached?: boolean;
    useAuth?: boolean;
}